Put tracing in its own file
